requirements.txt
fastapi
uvicorn
praw
python-dotenv